Add billing charge API

+++ src/notification.go
@@ -3,6 +3,6 @@ package main
 type NotificationMessage struct {
 	UserId      string `json:"userId"`
 	WorkspaceId string `json:"workspaceId"`
-	Amount      uint64 `json:"amount"`
+	Amount      uint32 `json:"amount"`
 	Timestamp   uint32 `json:"timestamp"`
 }
